requirements.txt
python-telegram-bot==22.5
python_dotenv==1.2.1
Django==5.2.8